Feedback from GMB.
Lifted from email:

> > Packet: An addressed item (usually P_Pkt_t, or P_Pkt_t if GMB’s change has
> > been accepted)
> It's a P_Pkt_t, just pick one of the two 😉

Fixed

> > There is no P_CNC_KILL. Applications can no longer stop the Mothership. A
> > normal device can send a packet to the supervisor, (via a (BEND, SUPR)
> > message), which causes the supervisor to call the void Super::end() API
> > method (documented in the Supervisor API section), which closes the
> > application gracefully.
>
> I'm not so sure about this. The KILL functionality was there partly to cope
> with a run-breaking issue (e.g. a failed assertion, crashing core,
> unresponsive supervisor, etc.) so that a device could indicate concisely that
> something really bad had gone wrong. While we don’t currently use this
> ability anywhere, it does still seem desireable/useful.

Put it back, because it's pretty easy to implement, though I am a little
suspicious of its usecase. I think in the case of a "something's gone wrong",
I'd try to send debug link messages and stop messages, which would more readily
identify if other cores experience the same problem.

> > There is no P_CNC_INIT. This opcode was used in the barrier-breaking packet
> > for normal devices. It is replaced by P_CNC_BARRIER.
>
> Just to confirm, this is essentially merging INIT and BARRIER?

Yes, this is correct.

> Also, I am not sure it is intentional, but you don't mention that an INSTR to
> a device solicits a return INSTR

Correct; I've just explicitly added a mention in the "notes" for the Tinsel
Command and Control section.

> Re. the commands in Table 1, have you reviewed how these are likely to
> interact with sBase/NameServer? There was some thought about tweaking the
> deploy so that sBase was involved as the TaskRecords have path information in
> them already.

My intention was for them to just not (for now), and to refactor this once
SBase is in. Added my intentions to the implementation plan section.

> > Can you think of any more? There are probably more we need, but we’re not
> > going to get them all now. Simply implementing a framework by which this
> > API can be extended should be enough.
> Didn't we have a list written down? There was something about Nameserver
> query methods and RTCL methods.

Somewhat shamefully we did have a list written down, but on rifling through
ADB's office I couldn't find them. I've added some extra notes on the
Supervisor API, but since I'm not exactly sure what people would want to do
with them, I've left them deliberately vague (for now)

> > Note that this API does not define methods for termination detection. It
> > could do (using a Softswitch-based heartbeats mechanism), but it might be
> > best to let sleeping dragons lie for now.
> This is one of the elephants in the room and a more Softswitch-integrated
> version of Heartbeats has to come with the v4 integration.

IIRC, the softswitch does not support this at present. I imagine the mechanism
would be simple, from the Mothership perspective, because it just waits for a
`P_CNC_HEARTBEAT` packet or so, and collects the results. The supervisor API
could define `bool Super::stop_if_all_heartbeats_received_or_something()`,
making it easy for the application writer to use the thing.

Alternatively, we can leave the sleeping dragon alone, lest it breathe fire and
melt our computers.

> Also, a few typos/formatting issues:
> * Footnote on page 4 splits onto page 5.
> * Page 9, "begin" at the end of the STOPPING entry should be "begun"

Fixed